Reworking process for creating an install image
The previous install method was to boot the Armbian image, set a
root password, and run the install over ssh.  I'm changing this to
preloading the install script on the Armbian image, so the install
happens on the first boot.  The idea is that once the install is
done, the system shuts down and then the image can be cloned onto
sdcards for new screens.